Switched from String[] to SQLiteDatabase
Added new DatabaseHandler class which the Words class uses to fetch
words for the main activity. Added new WordList class which just holds a
bunch of words to add to the words table on initial creation of the
database. Changed functionality of the Words to class to accomodate the
change from a String array to a DB.