DB changes
Added database file to assets folder so that a premade database will be
copied over onto the device when the app is loaded for the first time as
opposed to being created manually on start up. Added a "READ" field to
the database to flag when words have already been viewed by the user.
Changed the word lookup to grab random words as opposed to a specific
ID.